+++ src/main/java/com/example/demo/openapi/dto/AirQualityDTO.java
@@ -1,4 +1,4 @@
-package com.example.demo.openapi.model;
+package com.example.demo.openapi.dto;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.Data;
