+++ src/main/java/com/example/demo/openapi/dto/OceanQualityDTO.java
@@ -1,4 +1,4 @@
-package com.example.demo.openapi.model;
+package com.example.demo.openapi.dto;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
